Fix: don't crash when dragging outside the canvas
The app was crashing when clicking the drawing area and then dragging
outside of it.

To draw in the canvas, you first need to set the value of the pixel in
the grid array.
Then there's a function that colors cells based on the array.

The problem was updating the value of the array when clicking and drag-
ging the mouse outside the drawing area.
It was crashing due to attempting to update a value out of bounds.

Now, it just checks this before attempting to update the array.

There was no problem with the drawing function.